move theme functions out of main
