+++ js/main.js
@@ -7,46 +7,7 @@ import { stripMarkdown } from "./util";
 import { fetchLongNotes, fetchUserProfile } from "./fetch";
 import { displayLongNote } from "./article";
 import { displayLongNoteIndex } from "./article_index";
-
-// Function to set dark mode preference
-function setDarkModePreference(isDarkMode) {
-  localStorage.setItem("darkMode", isDarkMode);
-}
-
-// Function to get dark mode preference
-function getDarkModePreference() {
-  return JSON.parse(localStorage.getItem("darkMode"));
-}
-
-// Function to toggle dark mode
-function toggleDarkMode() {
-  const isDarkMode = getDarkModePreference();
-  const html = document.documentElement;
-  const darkModeToggle = document.getElementById("darkModeToggle");
-  
-  if (isDarkMode) {
-    html.setAttribute("data-theme", "dark");
-    darkModeToggle.querySelector("span").classList.add("translate-x-full");
-  } else {
-    html.removeAttribute("data-theme");
-    darkModeToggle.querySelector("span").classList.remove("translate-x-full");
-  }
-}
-//CHANGE THIS FUNCTION TO TOGGLE MORE THAN TWO THEMES
-
-
-// On page load, apply dark mode preference
-window.addEventListener("load", () => {
-  toggleDarkMode();
-});
-
-// Event listener for dark mode toggle button
-const darkModeToggle = document.getElementById("darkModeToggle");
-darkModeToggle.addEventListener("click", () => {
-  const isDarkMode = getDarkModePreference();
-  setDarkModePreference(!isDarkMode);
-  toggleDarkMode();
-});
+import "./theme"
 
 async function bootstrap() {
   try {
